tweaks dmenu; adds 'unmark' for latest i3
This rewrites `mark` to use dmenu. Since `mark` overwrites any previous
marks automatically, and the user might not remember which marks he has
added, it is handy to show them in a menu.

It also adds prompts to `goto` and `mark` dmenus.

Finally, the latest i3 build added an `unmark` command. This adds a
script to provide a dmenu for unmarks.